Move AppRoutes to components directory
